backup-python/requirement.txt
bcrypt==4.1.3
cffi==1.16.0
cryptography==42.0.8
future==1.0.0
netmiko==3.4.0
ntc_templates==5.1.0
packaging==20.9
paramiko==3.4.0
pycparser==2.22
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
python-decouple==3.8
routeros-ssh-connector==1.7.1
scp==0.15.0
six==1.16.0
tenacity==8.4.2
textfsm==1.1.3
urllib3==2.0.4